// src/types/article.types.ts
export type articles = {

    artile_id: number;
    articleby: string,
    published_img: string
    possition: string,
    img_url: any;
    title: string;
    published: string;
    para_one: string;
    img_url_two: string;
    para_two: string[];
    para_array: string[];
    img_list: string[];

}